# docs/swagger.yaml
basePath: /
definitions:
  db.Campaign:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      endDate:
        type: string
      id:
        type: integer
      maxParticipants:
        type: integer
      name:
        type: string
      startDate:
        type: string
      updated_at:
        type: string
    type: object
  db.CampaignUser:
    properties:
      campaign:
        $ref: '#/definitions/db.Campaign'
      campaignID:
        type: integer
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phoneNumber:
        type: string
      registrationMethod:
        type: string
      status:
        type: string
      subscriptionType:
        type: string
      updated_at:
        type: string
    type: object
  db.Voucher:
    properties:
      codePromotion:
        type: string
      created_at:
        type: string
      expiryDate:
        type: string
      id:
        type: integer
      status:
        type: string
      updated_at:
        type: string
      value:
        type: number
    type: object
  types.CampaignRegistrationInfo:
    type: object
  types.CampaignRequest:
    type: object
  util.EmptyResponse:
    type: object
host: localhost:8089
info:
  contact: {}
  title: Service API
  version: "1.0"
paths:
  /campaigns:
    get:
      consumes:
      - application/json
      description: This endpoint retrieves all campaign.
      parameters:
      - description: API key for authentication
        in: header
        name: x-api-key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Campaign retrieved successfully
          schema:
            $ref: '#/definitions/db.Campaign'
        "500":
          description: Error retrieving Campaign
          schema:
            $ref: '#/definitions/util.EmptyResponse'
      security:
      - ApiKeyAuth: []
      summary: Retrieve all campaign
      tags:
      - Campaign
    post:
      consumes:
      - application/json
      parameters:
      - description: API key for authentication
        in: header
        name: x-api-key
        required: true
        type: string
      - description: campaign information
        in: body
        name: campaign
        required: true
        schema:
          $ref: '#/definitions/types.CampaignRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Campaign created successfully
          schema:
            $ref: '#/definitions/util.EmptyResponse'
        "400":
          description: Invalid input or data
          schema:
            $ref: '#/definitions/util.EmptyResponse'
        "500":
          description: Failed to create campaign
          schema:
            $ref: '#/definitions/util.EmptyResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Campaign
      tags:
      - Campaign
  /campaigns/{id}:
    get:
      consumes:
      - application/json
      description: This endpoint retrieves a campaign by its ID.
      parameters:
      - description: API key for authentication
        in: header
        name: x-api-key
        required: true
        type: string
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Campaign retrieved successfully
          schema:
            $ref: '#/definitions/db.Campaign'
        "400":
          description: Invalid input or ID format
          schema:
            $ref: '#/definitions/util.EmptyResponse'
        "500":
          description: Error retrieving Campaign
          schema:
            $ref: '#/definitions/util.EmptyResponse'
      security:
      - ApiKeyAuth: []
      summary: Retrieve a campaign by ID
      tags:
      - Campaign
  /campaigns/register:
    get:
      consumes:
      - application/json
      description: This endpoint retrieves All campaignUser.
      parameters:
      - description: API key for authentication
        in: header
        name: x-api-key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: CampaignUser retrieved successfully
          schema:
            $ref: '#/definitions/db.CampaignUser'
        "500":
          description: Error retrieving CampaignUser
          schema:
            $ref: '#/definitions/util.EmptyResponse'
      security:
      - ApiKeyAuth: []
      summary: Retrieve All campaign user
      tags:
      - RegisterCampaign
    post:
      consumes:
      - application/json
      parameters:
      - description: API key for authentication
        in: header
        name: x-api-key
        required: true
        type: string
      - description: campaign information
        in: body
        name: registerInfo
        required: true
        schema:
          $ref: '#/definitions/types.CampaignRegistrationInfo'
      produces:
      - application/json
      responses:
        "200":
          description: Campaign registered successfully
          schema:
            $ref: '#/definitions/db.Voucher'
        "400":
          description: Invalid input or data
          schema:
            $ref: '#/definitions/util.EmptyResponse'
        "500":
          description: Failed to create campaign
          schema:
            $ref: '#/definitions/util.EmptyResponse'
      security:
      - ApiKeyAuth: []
      summary: Register Campaign
      tags:
      - RegisterCampaign
  /campaigns/register/{id}:
    get:
      consumes:
      - application/json
      description: This endpoint retrieves a campaignUser by its ID.
      parameters:
      - description: API key for authentication
        in: header
        name: x-api-key
        required: true
        type: string
      - description: CampaignUser ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: CampaignUser retrieved successfully
          schema:
            $ref: '#/definitions/db.CampaignUser'
        "400":
          description: Invalid input or ID format
          schema:
            $ref: '#/definitions/util.EmptyResponse'
        "500":
          description: Error retrieving CampaignUser
          schema:
            $ref: '#/definitions/util.EmptyResponse'
      security:
      - ApiKeyAuth: []
      summary: Retrieve a campaign User by ID
      tags:
      - RegisterCampaign
schemes:
- http
swagger: "2.0"